Implement threshold scan test. Many visual & UI changes on coldbox tab.

+++ scripts/load_udev.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
-
-sudo cp udev/99-ETL_Testing_GUI.rules /etc/udev/rules.d/
-sudo udevadm control --reload
-sudo udevadm trigger
+if [ -f udev/99-ETL_Testing_GUI.rules ] then
+    sudo cp udev/99-ETL_Testing_GUI.rules /etc/udev/rules.d/
+    sudo udevadm control --reload
+    sudo udevadm trigger
+fi
